+++ src/frequenz/client/base/channel.py
@@ -17,8 +17,16 @@ from grpc.aio import (
     secure_channel,
 )
 
-from .authentication import AuthenticationInterceptor, AuthenticationOptions
-from .signing import SigningInterceptor, SigningOptions
+from .authentication import (
+    AuthenticationInterceptorUnaryStream,
+    AuthenticationInterceptorUnaryUnary,
+    AuthenticationOptions,
+)
+from .signing import (
+    SigningInterceptorUnaryStream,
+    SigningInterceptorUnaryUnary,
+    SigningOptions,
+)
 
 
 @dataclasses.dataclass(frozen=True)
@@ -193,14 +201,16 @@ def parse_grpc_uri(
 
     interceptors: list[ClientInterceptor] = []
     if defaults.auth is not None:
-        interceptors.append(
-            AuthenticationInterceptor(options=defaults.auth)  # type: ignore[arg-type]
-        )
+        interceptors += [
+            AuthenticationInterceptorUnaryUnary(options=defaults.auth),  # type: ignore [list-item]
+            AuthenticationInterceptorUnaryStream(options=defaults.auth),  # type: ignore [list-item]
+        ]
 
     if defaults.sign is not None:
-        interceptors.append(
-            SigningInterceptor(options=defaults.sign)  # type: ignore[arg-type]
-        )
+        interceptors += [
+            SigningInterceptorUnaryUnary(options=defaults.sign),  # type: ignore [list-item]
+            SigningInterceptorUnaryStream(options=defaults.sign),  # type: ignore [list-item]
+        ]
 
     ssl = defaults.ssl.enabled if options.ssl is None else options.ssl
     if ssl:
@@ -278,13 +288,13 @@ def _parse_query_params(uri: str, query_string: str) -> _QueryParams:
     }
 
     if ssl is False:
-        erros = []
+        errors = []
         for opt_name, opt in ssl_opts.items():
             if opt is not None:
-                erros.append(opt_name)
-        if erros:
+                errors.append(opt_name)
+        if errors:
             raise ValueError(
-                f"Option(s) {', '.join(erros)} found in URI {uri!r}, but SSL is disabled",
+                f"Option(s) {', '.join(errors)} found in URI {uri!r}, but SSL is disabled",
             )
 
     keep_alive_option = options.pop("keep_alive", None)
@@ -298,13 +308,13 @@ def _parse_query_params(uri: str, query_string: str) -> _QueryParams:
     }
 
     if keep_alive is False:
-        erros = []
+        errors = []
         for opt_name, opt in keep_alive_opts.items():
             if opt is not None:
-                erros.append(opt_name)
-        if erros:
+                errors.append(opt_name)
+        if errors:
             raise ValueError(
-                f"Option(s) {', '.join(erros)} found in URI {uri!r}, but keep_alive is disabled",
+                f"Option(s) {', '.join(errors)} found in URI {uri!r}, but keep_alive is disabled",
             )
 
     if options: