+++ tests/test_authentication.py
@@ -5,24 +5,18 @@
 
 from unittest import mock
 
-from frequenz.client.base.authentication import (
-    AuthenticationInterceptor,
-    AuthenticationOptions,
-)
+from frequenz.client.base.authentication import _add_auth_header
 
 
 async def test_auth_interceptor() -> None:
     """Test that the Auth Interceptor adds the correct header."""
-    auth: AuthenticationOptions = AuthenticationOptions(api_key="my_key")
-    auth_interceptor: AuthenticationInterceptor = AuthenticationInterceptor(
-        options=auth
-    )
-
     metadata: dict[str, str] = {}
 
     client_call_details = mock.MagicMock(method="my_rpc")
     client_call_details.metadata = metadata
 
-    auth_interceptor.add_auth_header(client_call_details)
+    key = "my_key"
+
+    _add_auth_header(key, client_call_details)
 
     assert metadata["key"] == "my_key"